+++ app/Bpost/Queries/AllMunicipalitiesTest.php
@@ -18,10 +18,8 @@ final class AllMunicipalitiesTest extends TestCase
     #[Before]
     public function ensureNoFilePresent(): void
     {
-        if (File::exists($this->filePath)) 
-        {
-            File::delete($this->filePath);
-        }
+        
+        if (File::exists($this->filePath)) File::delete($this->filePath);
     }
 
     #[Test]
